+++ cmem_plugin_base/dataintegration/parameter/resource.py
@@ -2,11 +2,9 @@
 
 from typing import Any
 
-from cmem.cmempy.workspace.projects.resources import get_resources
-
+from cmem_plugin_base.dataintegration.client import get_client
 from cmem_plugin_base.dataintegration.context import PluginContext
 from cmem_plugin_base.dataintegration.types import Autocompletion, StringParameterType
-from cmem_plugin_base.dataintegration.utils import setup_cmempy_user_access
 
 
 class ResourceParameterType(StringParameterType):
@@ -23,12 +21,11 @@ class ResourceParameterType(StringParameterType):
         context: PluginContext,
     ) -> list[Autocompletion]:
         """Autocompletion request - Returns all results that match ALL provided query terms."""
-        setup_cmempy_user_access(context.user)
-        resources = get_resources(context.project_id)
+        resources = get_client(context).files.get_resources(context.project_id)
         result = [
             Autocompletion(
-                value=f"{_['fullPath']}",
-                label=f"{_['name']}",
+                value=f"{_.full_path}",
+                label=f"{_.name}",
             )
             for _ in resources
         ]